- replacing newly added VCFPathReader with Path constructors in VCFFileReader and deprecating the File versions
